Reverted to solo test in ray file after WandB issue fixed
